Fix/backend/resources

+++ backend/main/resources/user.py
@@ -9,34 +9,69 @@ from main.auth.decorators import role_required
 class User(Resource):
     @jwt_required()
     def get(self, id):
+        """
+        Handles GET requests to obtain a user.
+
+        Requires a valid JWT token for authentication. Returns a JSON representation
+        of the user if the user id matches the id in the JWT, otherwise returns a
+        JSON representation of the user with only the id, name, lastname, mail.
+        """
         user = db.session.query(UserModel).get_or_404(id)
         current_identity = get_jwt_identity()
-        print(current_identity)
         if current_identity == user.id:
             return user.to_json_complete()
         else:
             return user.to_json_short()
     
+    @jwt_required()
     @role_required(roles=['admin', 'user'])
     def delete(self, id):
+        """
+        Handles DELETE requests to delete a user.
+
+        Requires a valid JWT token for authentication and the role must be 'admin' or 'user'.
+        f the user id matches the id in the JWT or the role is 'admin', the request is completed.
+        The user is deleted from the database and a JSON representation of the user is
+        returned with a 204 status code. If an error occurs during processing, the transaction
+        is rolled back and an error message is returned with a 400 status code.
+        """
         user = db.session.query(UserModel).get_or_404(id)
+        current_identity = get_jwt_identity()
+        if current_identity != user.id and user.role != 'admin':
+            return {'error':'Unauthorized'}, 401
         try:
             db.session.delete(user)
             db.session.commit()
         except:
+            db.session.rollback()
             return {'error':'Incorrect data format'}, 400
         return user.to_json(), 204
     
     @jwt_required()
     def put(self, id):
+        """
+        Handles PUT requests to update a user.
+
+        Requires a valid JWT token for authentication and the user id must match the id in the JWT.
+        The request body must contain a JSON object with the fields to be modified. The fields that
+        can be modified are: name, lastname, alias, mail, phone, role (only for admin).
+        The user is updated in the database and a JSON representation of the user is returned with 
+        a 201 status code. If an error occurs during processing, the transaction is rolled back and
+        an error message is returned with a 400 status code.
+        """
         user = db.session.query(UserModel).get_or_404(id)
         data = request.get_json().items()
+        if user.id != get_jwt_identity():
+            return {'error':'Unauthorized'}, 401
+        for key, value in data:
+            if key == 'role' and user.role != 'admin':
+                continue
+            setattr(user, key, value)
         try:
-            for key, value in data:
-                setattr(user, key, value)
             db.session.add(user)
             db.session.commit()
         except:
+            db.session.rollback()
             return {'error':'Incorrect data format'}, 400
         return user.to_json() , 201 
 
@@ -44,9 +79,27 @@ class Users(Resource):
     @jwt_required()
     @role_required(roles=('admin', 'librarian'))
     def get(self):
-        # Default start page
+        """
+        Handles GET requests to obtain a list of users.
+
+        Requires a valid JWT token for authentication and one of the roles: admin, librarian.
+
+        The request body can contain the following parameters:
+        - page: the page number to retrieve.
+        - per_page: the number of items per page.
+        In case no parameters are provided, the default values are used: page=1 and per_page=10.
+
+        The response will contain a JSON object with the following structure:
+        - users: an array of user objects, each with the structure returned by the to_json() method.
+        - total: the total number of users.
+        - pages: the total number of pages.
+        - page: the current page number.
+
+        The response will return filtered users based on the request parameters.
+
+        If an error occurs during processing, the transaction is rolled back and an error message is returned with a 400 status code.
+        """
         page = 1
-        # Default pages quantity
         per_page = 10
         
         users = db.session.query(UserModel)
@@ -55,39 +108,28 @@ class Users(Resource):
         if request.args.get('per_page'):
             per_page = int(request.args.get('per_page'))
 
-        # Filters #
         if request.args.get('id'):
             users=users.filter(UserModel.id == request.args.get('id'))
-
         if request.args.get('name'):
             users=users.filter(UserModel.name.like('%'+request.args.get('name')+'%'))
-
         if request.args.get('lastname'):
             users=users.filter(UserModel.lastname.like('%'+request.args.get('lastname')+'%'))
-
         if request.args.get('mail'):
             users=users.filter(UserModel.mail.like('%'+request.args.get('mail')+'%'))
-
         if request.args.get('phone'):
             users=users.filter(UserModel.phone.like('%'+request.args.get('phone')+'%'))
-
         if request.args.get('rol'):
             users=users.filter(UserModel.rol.like('%'+request.args.get('rol')+'%'))
-
         if request.args.get('alias'):
             users=users.filter(UserModel.alias.like('%'+request.args.get('alias')+'%'))
-
         if request.args.get('rent'):
             users=users.filter(UserModel.rent.like('%'+request.args.get('rent')+'%'))
         
-        # Sort by #
-
         if request.args.get('sortby_name'):
             if request.args.get('sortby_name') == "desc":
                 users=users.order_by(desc(UserModel.name))
             else:
                 users=users.order_by(asc(UserModel.name))
-
         if request.args.get('sortby_lastname'):
             if request.args.get('sortby_lastname') == "desc":
                 users=users.order_by(desc(UserModel.lastname))
@@ -105,13 +147,13 @@ class Users(Resource):
     
     # This post method is to create a new user but without any restriction!!!
     # In production, this method should't be available.
-    # @role_required(roles=['admin']) # Uncomment this to activate restrictions
-    def post(self):
-        try:
-            user = UserModel.from_json(request.get_json())
-            db.session.add(user)
-            db.session.commit()
-        except:
-            db.session.rollback()
-            return {'error':'Incorrect data format'}, 400
-        return user.to_json(), 201
+    # # @role_required(roles=['admin']) # Uncomment this to activate restrictions
+    # def post(self):
+    #     user = UserModel.from_json(request.get_json())
+    #     try:
+    #         db.session.add(user)
+    #         db.session.commit()
+    #     except:
+    #         db.session.rollback()
+    #         return {'error':'Incorrect data format'}, 400
+    #     return user.to_json(), 201
